fixed addNewTax
broke because it was using $scope.taxTree before, but from a service it can't see that.

it's weird that deeper items (terms) are seen by the $digest when manipulated in their deeper parents, but when the parent of the items (tax) is a top-level scope item (taxTree) then $scope doesn't see the changes unless you run a $save.